requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pytesseract==0.3.10
Pillow==10.0.1
gTTS==2.3.2
opencv-python==4.8.1.78
Werkzeug==2.3.7
gunicorn==21.2.0
